requirements.txt
discord.py
black
humanfriendly
pytz
typing_extensions
tortoise-orm[asyncpg]